feat: use a function
